+++ src/test/java/uk/kukino/sgo/base/AdjacentTest.java
@@ -18,7 +18,7 @@ public class AdjacentTest
     public void size1Board()
     {
         short coord = Coord.parseToVal("A1");
-        long adjs = Adjacent.asVal(coord, (byte) 1);
+        int adjs = Adjacent.asVal(coord, (byte) 1);
 
         assertFalse(Adjacent.iterHasNext(adjs));
         assertThat(Adjacent.iterPosition(adjs), equalTo(Coord.INVALID));
@@ -27,7 +27,7 @@ public class AdjacentTest
     @Test
     public void happyPath()
     {
-        long adjs = Adjacent.asVal(c_E5, boardSize);
+        int adjs = Adjacent.asVal(c_E5, boardSize);
 
         assertThat(Adjacent.iterPosition(adjs), equalTo(Coord.parseToVal("E6")));
 
@@ -47,7 +47,7 @@ public class AdjacentTest
     @Test
     public void iteratesRightAmountAndStaysThere()
     {
-        long adjs = Adjacent.asVal(c_E5, boardSize);
+        int adjs = Adjacent.asVal(c_E5, boardSize);
 
         int count = 0;
         while (Adjacent.iterHasNext(adjs))
@@ -63,7 +63,7 @@ public class AdjacentTest
     @Test
     public void itCanBeReset()
     {
-        long adjs = Adjacent.asVal(c_E5, boardSize);
+        int adjs = Adjacent.asVal(c_E5, boardSize);
 
         adjs = assertPositions(adjs, "E6", "F5", "E4", "D5");
         assertThat(Adjacent.iterPosition(adjs), equalTo(Coord.INVALID));
@@ -76,7 +76,7 @@ public class AdjacentTest
     @Test
     public void itCanBeInteractivelyUnset()
     {
-        long adjs = Adjacent.asVal(c_E5, boardSize);
+        int adjs = Adjacent.asVal(c_E5, boardSize);
         assertPositions(adjs, "E6", "F5", "E4", "D5");
 
         adjs = Adjacent.iterReset(adjs);
@@ -94,7 +94,7 @@ public class AdjacentTest
     public void unsettingDoNotMoveIteratorPosition()
     {
         // we don't want to do two operations in one
-        long adjs = Adjacent.asVal(c_E5, boardSize);
+        int adjs = Adjacent.asVal(c_E5, boardSize);
         assertThat(Adjacent.valToStr(adjs), equalTo("{E5 NESW #0->E6}"));
 
         adjs = Adjacent.iterUnset(adjs);
@@ -104,56 +104,56 @@ public class AdjacentTest
     @Test
     public void topLeft()
     {
-        long adjs = Adjacent.asVal(Coord.parseToVal("A9"), boardSize);
+        int adjs = Adjacent.asVal(Coord.parseToVal("A9"), boardSize);
         assertPositions(adjs, "B9", "A8");
     }
 
     @Test
     public void top()
     {
-        long adjs = Adjacent.asVal(Coord.parseToVal("E9"), boardSize);
+        int adjs = Adjacent.asVal(Coord.parseToVal("E9"), boardSize);
         assertPositions(adjs, "F9", "E8", "D9");
     }
 
     @Test
     public void topRight()
     {
-        long adjs = Adjacent.asVal(Coord.parseToVal("J9"), boardSize);
+        int adjs = Adjacent.asVal(Coord.parseToVal("J9"), boardSize);
         assertPositions(adjs, "J8", "H9");
     }
 
     @Test
     public void right()
     {
-        long adjs = Adjacent.asVal(Coord.parseToVal("J5"), boardSize);
+        int adjs = Adjacent.asVal(Coord.parseToVal("J5"), boardSize);
         assertPositions(adjs, "J6", "J4", "H5");
     }
 
     @Test
     public void bottomRight()
     {
-        long adjs = Adjacent.asVal(Coord.parseToVal("J1"), boardSize);
+        int adjs = Adjacent.asVal(Coord.parseToVal("J1"), boardSize);
         assertPositions(adjs, "J2", "H1");
     }
 
     @Test
     public void bottom()
     {
-        long adjs = Adjacent.asVal(Coord.parseToVal("E1"), boardSize);
+        int adjs = Adjacent.asVal(Coord.parseToVal("E1"), boardSize);
         assertPositions(adjs, "E2", "F1", "D1");
     }
 
     @Test
     public void bottomLeft()
     {
-        long adjs = Adjacent.asVal(Coord.parseToVal("A1"), boardSize);
+        int adjs = Adjacent.asVal(Coord.parseToVal("A1"), boardSize);
         assertPositions(adjs, "A2", "B1");
     }
 
     @Test
     public void left()
     {
-        long adjs = Adjacent.asVal(Coord.parseToVal("A5"), boardSize);
+        int adjs = Adjacent.asVal(Coord.parseToVal("A5"), boardSize);
         assertPositions(adjs, "A6", "B5", "A4");
     }
 
@@ -165,7 +165,7 @@ public class AdjacentTest
     }
     // util
 
-    public static long assertPositions(long adjs, String... positions)
+    public static int assertPositions(int adjs, String... positions)
     {
         int posIdx = 0;
         while (Adjacent.iterHasNext(adjs))
